fix: ensure user plugin's resolvedId() 'meta' reply isn't lost
The normalizeUrl helper calls user plugins's resolveId() twice.

* Once with the URL string containing a query suffix.  Here it
  _ignores_ the meta reply.

* Again, with the URL no longer containing said suffix.  Here it
  doesn't ignore the meta reply.  It is stored in the moduleGraph
  node.

As can be seen, if the user's plugin meta reply depends on the query
suffix, it won't be registered correctly in the module graph.

This change makes it so that the second resolution call is actually
passed the query suffix.  Then when updating the moduleGraph for the
correct queryless URL, the meta reply obtained before is preserved.

This makes Vite's dev server match Rollup's plugin-calling behaviour
and fixes #6766

Fix: #6766